chore: remove unused methods
